refactor (apiclient): Rename client to API client and other changes
