+++ sdks/js/packages/core/ApiClient/index.ts
@@ -1,0 +1,4 @@
+import { V1Beta1 } from './V1Beta1';
+
+export * from './data-contracts';
+export const ApiClient = V1Beta1;
