--- sdks/js/packages/core/client/index.ts
@@ -1,2 +1,0 @@
-export * from './data-contracts';
-export * from './V1Beta1';
