+++ sdks/js/packages/core/src/index.ts
@@ -1,5 +1,5 @@
-import { V1Beta1 } from '../client/V1Beta1';
+import { V1Beta1 } from '../ApiClient/V1Beta1';
 
-export * from '../client/V1Beta1';
-export * from '../client/data-contracts';
+export * from '../ApiClient/V1Beta1';
+export * from '../ApiClient/data-contracts';
 export const FrontierClient = V1Beta1;
