+++ sdks/js/packages/sdk-demo/src/api/frontier.ts
@@ -1,0 +1,11 @@
+import { ApiClient } from '../../../core/ApiClient/dist'
+import config from '@/config/frontier';
+
+const client = new ApiClient({
+    baseUrl: config.endpoint,
+    baseApiParams: {
+        credentials: 'include'
+    }
+});
+
+export default client
